indexer/requirements.txt
google-cloud-speech==0.36.0
google-cloud-storage==1.12.0
pydub==0.22.1
requests
